+++ CHANGELOG.md
@@ -8,6 +8,19 @@ and this project adheres to
 
 ## [Unreleased]
 
+## 2.0.0 - 2022-10-11
+
+### Changed
+
+- Renamed config and environment variables to align with use
+  - CLIENT_URL -> LDAP_URL
+  - CLIENT_DOMAIN -> BASE_DN
+  - CLIENT_USERNAME -> USERNAME
+  - CLIENT_PASSWORD -> PASSWORD
+- Updated jupiterone.md and development.md docs
+- Updated `ldapts` package
+- Changed LDAP user query
+
 ## 1.0.0 - 2022-04-26
 
 ### Added